refactor: split main.go and session.go into focused files
Pure file moves, no logic changes. Splits within package main:

- cli_dispatch.go: commandDispatch map, printHelp, printConfigHelp, printVersion
- cli_serve.go: runServe + serverConfig + flag parsing + bindListener +
  runIdleTimeoutChecker + applySessionOverrides + the daemon-boot path
- cli_install.go: runInstall + installIntegration + integrationMap +
  destFor / installOneFile / printUniqueHints / globalDestKind helpers
- session_focus.go: Focus / FocusKind / DiffScope types + visibleInFocus +
  stampWithFocus + SetFocus + scoped-view computation
  (cachedAvailableScopes, GetSessionInfoScoped, GetFileDiffSnapshotScoped,
  loadScopedFileState, computeScopedDiffHunks)
- session_write.go: WriteFiles + scheduleWrite debounce + snapshotForWrite +
  buildCritJSON + merge* helpers + handleExternalDeletion +
  clearAllCommentData + critJSONIsEmpty + mergeExternalCritJSON

main.go drops from 3151 -> 2150 LOC. session.go drops from 3189 -> 1932 LOC.
