+++ main.go
@@ -1,25 +1,18 @@
 package main
 
 import (
-	"context"
 	"embed"
 	"encoding/json"
 	"errors"
 	"flag"
 	"fmt"
 	"io"
-	"log"
-	"net"
 	"net/http"
 	"os"
-	"os/exec"
 	"os/signal"
 	"path/filepath"
-	"runtime"
-	"sort"
 	"strconv"
 	"strings"
-	"sync"
 	"syscall"
 	"time"
 
@@ -37,33 +30,6 @@ var (
 	commit  = "unknown"
 	date    = "unknown"
 )
-
-// commandDispatch maps subcommand names to handler functions.
-var commandDispatch = map[string]func([]string){
-	"help":      func([]string) { printHelp() },
-	"--help":    func([]string) { printHelp() },
-	"-h":        func([]string) { printHelp() },
-	"--version": func([]string) { printVersion() },
-	"-v":        func([]string) { printVersion() },
-	"share":     runShare,
-	"fetch":     runFetch,
-	"unpublish": runUnpublish,
-	"install":   runInstall,
-	"config":    runConfig,
-	"check":     func([]string) { runCheck() },
-	"pr":        runPR,
-	"pull":      runPull,
-	"push":      runPush,
-	"comment":   runComment,
-	"review":    runReview,
-	"plan":      runPlan,
-	"plan-hook": func([]string) { runPlanHook() },
-	"auth":      runAuth,
-	"stop":      runStop,
-	"status":    runStatus,
-	"cleanup":   runCleanup,
-	"_serve":    runServe,
-}
 
 func main() {
 	if len(os.Args) < 2 {
@@ -430,48 +396,6 @@ func runUnpublish(args []string) {
 	}
 
 	fmt.Println("Review unpublished.")
-}
-
-func runInstall(args []string) {
-	if len(args) < 1 {
-		fmt.Fprintln(os.Stderr, "Usage: crit install <agent>")
-		fmt.Fprintln(os.Stderr, "")
-		fmt.Fprintln(os.Stderr, "Available agents:")
-		for _, a := range availableIntegrations() {
-			fmt.Fprintf(os.Stderr, "  %s\n", a)
-		}
-		fmt.Fprintln(os.Stderr, "  all")
-		os.Exit(1)
-	}
-
-	force := false
-	for _, arg := range args[1:] {
-		if arg == "--force" || arg == "-f" {
-			force = true
-		}
-	}
-
-	target := args[0]
-	if target == "all" {
-		cwd := mustGetwd()
-		home, _ := os.UserHomeDir()
-		global := isGlobalInstall(cwd, home)
-		for _, name := range availableIntegrations() {
-			if name == "windsurf" && global {
-				fmt.Fprintln(os.Stderr, "  Skipped: windsurf (no global install supported — run from a project)")
-				continue
-			}
-			if err := installIntegration(name, force); err != nil {
-				fmt.Fprintf(os.Stderr, "  Failed: %s: %v\n", name, err)
-				continue
-			}
-		}
-		return
-	}
-	if err := installIntegration(target, force); err != nil {
-		fmt.Fprintln(os.Stderr, "Error: "+err.Error())
-		os.Exit(1)
-	}
 }
 
 func runConfig(args []string) {
@@ -1910,562 +1834,6 @@ func mustGetwd() string {
 	return wd
 }
 
-// serverConfig holds the resolved configuration for running the server.
-// It combines CLI flags, environment variables, and config file settings.
-type serverConfig struct {
-	port               int
-	noOpen             bool
-	quiet              bool
-	shareURL           string
-	authToken          string
-	outputDir          string
-	author             string
-	baseBranch         string // --base-branch override for diff base
-	ignorePatterns     []string
-	files              []string // explicit file arguments (empty = git mode)
-	noIntegrationCheck bool
-	noUpdateCheck      bool
-	agentCmd           string
-	planDir            string // managed storage directory for plan mode
-	planName           string // display name for plan content
-	reviewPath         string // centralized review file path (~/.crit/reviews/<key>.json)
-	vcsOverride        string // "git", "sl"/"sapling", or "" for auto-detect
-	cfg                Config // full resolved config for the settings panel
-
-	// focus is populated by resolveFocus when --pr or --range is set;
-	// nil means "default" (working-tree, derived inside the session).
-	focus *Focus
-
-	// remoteFiles enables API-based file content reads (gh api repos/.../contents/)
-	// when in PR/range focus, bypassing the local-fetch + git show path. Diff and
-	// changed-file lists still use local git.
-	remoteFiles bool
-
-	// workingTree forces working-tree mode regardless of stack auto-detection.
-	// Set by --working-tree.
-	workingTree bool
-}
-
-// serverFlagSet holds the parsed flag values before config resolution.
-type serverFlagSet struct {
-	port        int
-	noOpen      bool
-	showVersion bool
-	shareURL    string
-	outputDir   string
-	quiet       bool
-	noIgnore    bool
-	baseBranch  string
-	vcsOverride string
-	planDir     string
-	planName    string
-	fileArgs    []string
-
-	// PR-scoped / commit-range review (issue #300).
-	prSpec    string // --pr <num|url>
-	rangeSpec string // --range <baseSHA>..<headSHA>
-	scopeSpec string // --scope layer | full-stack
-
-	// remoteFiles is the parsed --remote flag. When true, file content reads
-	// in PR/range mode go through `gh api` instead of local git.
-	remoteFiles bool
-
-	// workingTree is the parsed --working-tree flag. When true, the boot
-	// path skips stacked-PR / local-stack auto-detection.
-	workingTree bool
-}
-
-func parseServerFlags(args []string) serverFlagSet {
-	fs := flag.NewFlagSet("crit", flag.ExitOnError)
-	port := fs.Int("port", 0, "Port to listen on (default: random available port)")
-	fs.IntVar(port, "p", 0, "Port to listen on (shorthand)")
-	noOpen := fs.Bool("no-open", false, "Don't auto-open browser")
-	showVersion := fs.Bool("version", false, "Print version and exit")
-	fs.BoolVar(showVersion, "v", false, "Print version and exit (shorthand)")
-	shareURL := fs.String("share-url", "", "Base URL of hosted Crit service for sharing reviews (overrides CRIT_SHARE_URL env var)")
-	outputDir := fs.String("output", "", "Output directory for review file (default: repo root or file directory)")
-	fs.StringVar(outputDir, "o", "", "Output directory for review file (shorthand)")
-	quiet := fs.Bool("quiet", false, "Suppress status output")
-	fs.BoolVar(quiet, "q", false, "Suppress status output (shorthand)")
-	noIgnore := fs.Bool("no-ignore", false, "Disable all ignore patterns from config files")
-	baseBranch := fs.String("base-branch", "", "Base branch to diff against (overrides auto-detection)")
-	vcsFlag := fs.String("vcs", "", "VCS backend to use: git, sl/sapling (default: auto-detect)")
-	planDir := fs.String("plan-dir", "", "")
-	planName := fs.String("name", "", "")
-	prSpec := fs.String("pr", "", "Review a specific PR by number or URL (e.g. 295 or https://github.com/o/r/pull/295)")
-	rangeSpec := fs.String("range", "", "Review a commit range, base..head (e.g. abc1234..def5678)")
-	scopeSpec := fs.String("scope", "", "Diff scope when reviewing a PR: layer (default) or full-stack")
-	remoteFiles := fs.Bool("remote", false, "Read PR file content via GitHub API instead of local git (avoids `git fetch`; requires gh)")
-	workingTree := fs.Bool("working-tree", false, "Force working-tree mode (skip auto-detection of stacked PR / branch)")
-	fs.Usage = func() {
-		printHelp()
-	}
-	fs.Parse(args)
-
-	return serverFlagSet{
-		port:        *port,
-		noOpen:      *noOpen,
-		showVersion: *showVersion,
-		shareURL:    *shareURL,
-		outputDir:   *outputDir,
-		quiet:       *quiet,
-		noIgnore:    *noIgnore,
-		baseBranch:  *baseBranch,
-		vcsOverride: *vcsFlag,
-		planDir:     *planDir,
-		planName:    *planName,
-		fileArgs:    fs.Args(),
-		prSpec:      *prSpec,
-		rangeSpec:   *rangeSpec,
-		scopeSpec:   *scopeSpec,
-		remoteFiles: *remoteFiles,
-		workingTree: *workingTree,
-	}
-}
-
-func resolvePort(flagPort, cfgPort int) int {
-	if flagPort != 0 {
-		return flagPort
-	}
-	if envPort := os.Getenv("CRIT_PORT"); envPort != "" {
-		if p, err := strconv.Atoi(envPort); err == nil {
-			return p
-		}
-	}
-	return cfgPort
-}
-
-func applyConfigDefaults(sf *serverFlagSet, cfg Config) {
-	sf.port = resolvePort(sf.port, cfg.Port)
-	if !sf.noOpen && cfg.NoOpen {
-		sf.noOpen = true
-	}
-	sf.shareURL = resolveShareURL(sf.shareURL, cfg, "")
-	if !sf.quiet && cfg.Quiet {
-		sf.quiet = true
-	}
-	if sf.outputDir == "" && cfg.Output != "" {
-		sf.outputDir = cfg.Output
-	}
-	if sf.baseBranch == "" && cfg.BaseBranch != "" {
-		sf.baseBranch = cfg.BaseBranch
-	}
-	if sf.baseBranch != "" {
-		setDefaultBranchOverride(sf.baseBranch)
-	}
-}
-
-// resolveServerConfig parses flags, loads config files, and resolves the
-// final server configuration from all sources (CLI > env > config > defaults).
-// Returns nil when the command should exit early (e.g. --version).
-//
-//nolint:unparam // error return is future-proofing for config validation
-func resolveServerConfig(args []string) (*serverConfig, error) {
-	sf := parseServerFlags(args)
-
-	if sf.showVersion {
-		printVersion()
-		return nil, nil
-	}
-
-	configDir := ""
-	vcs := DetectVCS(sf.vcsOverride)
-	repoRoot := ""
-	if vcs != nil {
-		configDir, _ = vcs.RepoRoot()
-		repoRoot = configDir
-	}
-	if configDir == "" {
-		configDir = mustGetwd()
-	}
-	cfg := LoadConfig(configDir)
-
-	applyConfigDefaults(&sf, cfg)
-
-	var ignorePatterns []string
-	if !sf.noIgnore {
-		ignorePatterns = cfg.IgnorePatterns
-	}
-
-	// Resolve --pr / --range / --scope into a Focus. nil = working-tree default.
-	focus, err := resolveFocus(sf.prSpec, sf.rangeSpec, sf.scopeSpec, sf.remoteFiles, vcs, repoRoot)
-	if err != nil {
-		return nil, err
-	}
-
-	// --remote only takes effect in PR/range mode. Warn but don't fail.
-	if sf.remoteFiles && focus == nil {
-		fmt.Fprintln(os.Stderr, "Warning: --remote has no effect without --pr or --range; ignoring")
-	}
-
-	return &serverConfig{
-		port:               sf.port,
-		noOpen:             sf.noOpen,
-		quiet:              sf.quiet,
-		shareURL:           sf.shareURL,
-		authToken:          cfg.AuthToken,
-		outputDir:          sf.outputDir,
-		author:             cfg.Author,
-		baseBranch:         sf.baseBranch,
-		ignorePatterns:     ignorePatterns,
-		noIntegrationCheck: cfg.NoIntegrationCheck,
-		noUpdateCheck:      cfg.NoUpdateCheck,
-		agentCmd:           cfg.AgentCmd,
-		files:              sf.fileArgs,
-		planDir:            sf.planDir,
-		planName:           sf.planName,
-		vcsOverride:        resolveVCSOverride(sf.vcsOverride, cfg.VCS),
-		cfg:                cfg,
-		focus:              focus,
-		remoteFiles:        sf.remoteFiles,
-		workingTree:        sf.workingTree,
-	}, nil
-}
-
-// resolveVCSOverride returns the effective VCS override.
-// --vcs flag takes precedence over config "vcs" field.
-func resolveVCSOverride(flag, config string) string {
-	if flag != "" {
-		return flag
-	}
-	return config
-}
-
-func createSession(sc *serverConfig) (*Session, error) {
-	var session *Session
-	var err error
-	if len(sc.files) == 0 {
-		vcs := DetectVCS(sc.vcsOverride)
-		if vcs == nil {
-			return nil, fmt.Errorf("not in a version-controlled repository and no files specified")
-		}
-		if sc.baseBranch != "" {
-			vcs.SetDefaultBranchOverride(sc.baseBranch)
-		}
-		session, err = NewSessionFromVCS(vcs, sc.ignorePatterns)
-	} else {
-		session, err = NewSessionFromFiles(sc.files, sc.ignorePatterns)
-	}
-	if err != nil {
-		return nil, err
-	}
-	// Apply --base-branch override to the session's VCS instance. This covers
-	// files mode where resolveGitContext creates a fresh VCS that doesn't have
-	// the override yet. For Sapling, the instance-level field must be set.
-	if sc.baseBranch != "" && session.VCS != nil {
-		session.VCS.SetDefaultBranchOverride(sc.baseBranch)
-	}
-	// Set ReviewFilePath before loadCritJSON so it reads from the centralized
-	// review file.
-	if sc.reviewPath != "" {
-		session.ReviewFilePath = sc.reviewPath
-		session.loadCritJSON()
-	}
-	return session, nil
-}
-
-func applySessionOverrides(session *Session, sc *serverConfig) {
-	if sc.planDir != "" {
-		applyPlanOverrides(session, sc.planDir, sc.planName)
-		for _, f := range session.Files {
-			f.Comments = []Comment{}
-		}
-		session.reviewComments = nil
-		session.loadCritJSON()
-	}
-	if sc.outputDir != "" {
-		abs, _ := filepath.Abs(sc.outputDir)
-		session.OutputDir = abs
-	}
-	// Auto-detect stacked PR / local branch stack when no explicit focus
-	// was requested. Best-effort — any failure falls through to working-tree
-	// mode. Skipped in file mode or by --working-tree.
-	if sc.focus == nil && !sc.workingTree && len(sc.files) == 0 {
-		if detected := autoDetectStackedFocus(session.VCS, session.RepoRoot); detected != nil {
-			sc.focus = detected
-		}
-	}
-	// Apply --pr / --range focus, if requested. SetFocus rebuilds the file
-	// list from the SHA range and persists ActiveDiffScope; failure leaves
-	// the working-tree state intact and reports via stderr.
-	if sc.focus != nil {
-		// Set RemoteFiles BEFORE SetFocus so the focus rebuild's file content
-		// reads route through the API path instead of local git.
-		session.RemoteFiles = sc.remoteFiles
-		if err := session.SetFocus(*sc.focus); err != nil {
-			fmt.Fprintf(os.Stderr, "Warning: failed to apply focus: %v\n", err)
-			return
-		}
-	}
-}
-
-func bindListener(port int) (net.Listener, error) {
-	var listener net.Listener
-	var err error
-	for attempt := 0; attempt < 3; attempt++ {
-		listener, err = net.Listen("tcp", fmt.Sprintf("127.0.0.1:%d", port))
-		if err == nil {
-			return listener, nil
-		}
-		if port == 0 {
-			break
-		}
-		time.Sleep(500 * time.Millisecond)
-	}
-	return nil, err
-}
-
-func serveSessionKey(sc *serverConfig) string {
-	cwd, _ := resolvedCWD()
-	if sc.planDir != "" {
-		return planSessionKey(cwd, sc.planName)
-	}
-	branch := ""
-	if vcs := DetectVCS(sc.vcsOverride); vcs != nil {
-		branch = vcs.CurrentBranch()
-	}
-	return sessionKey(cwd, branch, focusKeyArgs(sc))
-}
-
-func checkStaleIntegrations(sc *serverConfig, srv *Server, cwd string) {
-	if sc.noIntegrationCheck || os.Getenv("CRIT_NO_INTEGRATION_CHECK") != "" {
-		return
-	}
-	if home, err := os.UserHomeDir(); err == nil {
-		stale := checkInstalledIntegrations(cwd, home)
-		srv.staleIntegrations = stale
-		if len(stale) > 0 {
-			go printStaleWarnings(stale)
-		}
-	}
-}
-
-func runIdleTimeoutChecker(ctx context.Context, stop context.CancelFunc, idleMu *sync.Mutex, lastActivity *time.Time) {
-	const idleTimeout = 1 * time.Hour
-	ticker := time.NewTicker(5 * time.Minute)
-	defer ticker.Stop()
-	for {
-		select {
-		case <-ticker.C:
-			idleMu.Lock()
-			idle := time.Since(*lastActivity)
-			idleMu.Unlock()
-			if idle >= idleTimeout {
-				stop()
-				return
-			}
-		case <-ctx.Done():
-			return
-		}
-	}
-}
-
-func runServe(args []string) {
-	pipe := openReadyPipe()
-
-	sc, err := resolveServerConfig(args)
-	if err != nil {
-		daemonFatal(pipe, "Error: %v", err)
-	}
-	if sc == nil {
-		return
-	}
-	sc.quiet = true
-
-	listener, err := bindListener(sc.port)
-	if err != nil {
-		daemonFatal(pipe, "Error starting server: %v", err)
-	}
-	addr := listener.Addr().(*net.TCPAddr)
-
-	srv, err := NewServer(nil, frontendFS, sc.shareURL, sc.authToken, sc.author, version, addr.Port, sc.agentCmd)
-	if err != nil {
-		daemonFatal(pipe, "Error creating server: %v", err)
-	}
-
-	// Set config-dependent fields for the settings panel
-	srv.cfg = sc.cfg
-	cwd, _ := resolvedCWD()
-	srv.projectDir = cwd
-	if home, err := os.UserHomeDir(); err == nil {
-		srv.homeDir = home
-	}
-	key := serveSessionKey(sc)
-	branch := ""
-	if vcs := DetectVCS(sc.vcsOverride); vcs != nil {
-		branch = vcs.CurrentBranch()
-	}
-	if sc.outputDir != "" {
-		abs, _ := filepath.Abs(sc.outputDir)
-		sc.reviewPath = filepath.Join(abs, ".crit.json")
-	} else {
-		sc.reviewPath, _ = reviewFilePath(key)
-	}
-	srv.reviewPath = sc.reviewPath
-	srv.cliArgs = sc.files
-	if err := writeSessionFile(key, sessionEntry{
-		PID:        os.Getpid(),
-		Port:       addr.Port,
-		CWD:        cwd,
-		Args:       sc.files,
-		Branch:     branch,
-		ReviewPath: sc.reviewPath,
-		StartedAt:  time.Now().UTC().Format(time.RFC3339),
-	}); err != nil {
-		daemonFatal(pipe, "Error writing session file: %v", err)
-	}
-
-	var idleMu sync.Mutex
-	lastActivity := time.Now()
-	resetActivity := func() {
-		idleMu.Lock()
-		lastActivity = time.Now()
-		idleMu.Unlock()
-	}
-
-	httpServer := &http.Server{
-		Handler: http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			resetActivity()
-			srv.ServeHTTP(w, r)
-		}),
-		ReadTimeout: 15 * time.Second,
-		IdleTimeout: 60 * time.Second,
-	}
-
-	ctx, stop := signal.NotifyContext(context.Background(), syscall.SIGINT, syscall.SIGTERM, syscall.SIGHUP)
-	defer stop()
-
-	// Wire the shutdown ctx into the server so background goroutines (agent
-	// runner, etc.) can be cancelled on SIGINT/SIGTERM/idle-timeout instead of
-	// orphaning subprocesses and racing with WriteFiles.
-	srv.SetShutdownCtx(ctx)
-
-	go func() {
-		if err := httpServer.Serve(listener); err != http.ErrServerClosed {
-			log.Printf("Server error: %v", err)
-			stop()
-		}
-	}()
-
-	signalReadiness(pipe, addr.Port)
-
-	if !sc.noOpen {
-		go openBrowser(fmt.Sprintf("http://localhost:%d", addr.Port))
-	}
-
-	// Prime the open-PR cache in the background. `gh pr list` can take
-	// 2-5s on large orgs and the picker waits on it; running this during
-	// boot means the first /api/picker call lands on a warm cache instead
-	// of paying the network cost while the user watches the page render.
-	// Best-effort — failures (no gh, no remote, file mode) are silently
-	// dropped; the picker handler still degrades gracefully. Tied to the
-	// daemon's shutdown ctx so a Ctrl+C during boot terminates the gh
-	// subprocess instead of orphaning it.
-	if srv.prList != nil {
-		go func() { _, _ = srv.prList.getCtx(ctx) }()
-	}
-
-	go runIdleTimeoutChecker(ctx, stop, &idleMu, &lastActivity)
-
-	type sessionResult struct {
-		session *Session
-		err     error
-	}
-	ch := make(chan sessionResult, 1)
-	// NOTE: On timeout, the createSession goroutine will leak until its git
-	// operations finish (no context is threaded into the git calls). This is
-	// acceptable because the timeout path sets initErr, which triggers a full
-	// server shutdown and process exit shortly after, cleaning up all goroutines.
-	go func() {
-		s, err := createSession(sc)
-		ch <- sessionResult{s, err}
-	}()
-
-	var session *Session
-	var initErr error
-	select {
-	case res := <-ch:
-		session, initErr = res.session, res.err
-	case <-time.After(2 * time.Minute):
-		initErr = fmt.Errorf("session initialization timed out after 2 minutes")
-	}
-	if initErr != nil {
-		log.Printf("Error: %v", initErr)
-		srv.SetInitErr(initErr)
-		// Trigger immediate shutdown instead of waiting for SIGINT or the
-		// idle timeout. Without this the daemon would sit in 503-land for up
-		// to an hour after a failed init, burning a port and a process slot.
-		stop()
-		<-ctx.Done()
-		removeSessionFile(key)
-		shutCtx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
-		defer cancel()
-		_ = httpServer.Shutdown(shutCtx)
-		return
-	}
-	applySessionOverrides(session, sc)
-	session.CLIArgs = sc.files
-
-	checkStaleIntegrations(sc, srv, cwd)
-
-	if !sc.noUpdateCheck && os.Getenv("CRIT_NO_UPDATE_CHECK") == "" {
-		go srv.CheckForUpdates()
-	}
-	srv.SetSession(session)
-
-	if session.Mode == "git" {
-		go func() {
-			if prInfo := detectPRInfo(); prInfo != nil {
-				srv.SetPRInfo(prInfo)
-			}
-		}()
-	}
-
-	watchStop := make(chan struct{})
-	go session.Watch(watchStop)
-
-	<-ctx.Done()
-	close(watchStop)
-
-	removeSessionFile(key)
-
-	// Order matters here:
-	//   1. session.Shutdown()    — fires the final SSE "server-shutdown" event
-	//                              while clients are still connected.
-	//   2. httpServer.Shutdown() — stops accepting new conns and waits for
-	//                              in-flight handlers to return. This is what
-	//                              gates s.bgWG.Add(1): handleAgentRequest
-	//                              calls Add synchronously inside the handler
-	//                              before responding 202. Once Shutdown
-	//                              returns, no new Add() calls can race with
-	//                              the WaitBackground below — sync.WaitGroup
-	//                              would otherwise panic on Add-during-Wait.
-	//   3. WaitBackground        — drain spawned agent runners so their
-	//                              replies land before WriteFiles persists.
-	//                              Capped at 30s: a wedged agent loses its
-	//                              reply rather than hanging the daemon. The
-	//                              agent ctx is parented on the shutdown ctx
-	//                              (already Done() above), so subprocesses
-	//                              are being killed concurrently — most
-	//                              cases drain in milliseconds.
-	//   4. WriteFiles            — persist final review state.
-	session.Shutdown()
-
-	shutCtx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
-	defer cancel()
-	_ = httpServer.Shutdown(shutCtx)
-
-	if !srv.WaitBackground(30 * time.Second) {
-		log.Printf("Warning: background goroutines did not drain within 30s; proceeding with shutdown")
-	}
-
-	session.WriteFiles()
-
-	if session.ReviewFilePath != "" {
-		fmt.Fprintf(os.Stderr, "Review file: %s\n", session.ReviewFilePath)
-	}
-}
-
 func runStatus(args []string) {
 	jsonOutput := false
 	for _, arg := range args {
@@ -2779,373 +2147,4 @@ func deleteStaleReviews(stale []staleReview) int {
 		}
 	}
 	return deleted
-}
-
-func printHelp() {
-	fmt.Fprintf(os.Stderr, `crit — inline code review for AI agent workflows
-
-Usage:
-  crit                                       Auto-detect changed files via git
-  crit <file|dir> [...]                      Review specific files or directories
-  crit --pr <num|url>                        Review a GitHub pull request (range mode)
-  crit pr <num|url>                          Review a GitHub pull request (alias for --pr)
-  crit --range <baseSHA>..<headSHA>          Review a commit range (range mode)
-  crit stop [files...]                       Stop the daemon for current directory (and args)
-  crit stop --all                            Stop all daemons for current directory
-  crit comment <path>:<line[-end]> <body>    Add a review comment
-  crit comment --reply-to <id> [--resolve] [--author <name>] <body>  Reply to a comment
-  crit comment --json [--author <name>] [--output <dir>]    Read comments from stdin as JSON
-  crit comment --clear                       Remove all comments from the review file
-  crit share <file> [file...]                Share files to crit-web and print the URL
-  crit fetch [--output <dir>]               Fetch comments from crit-web into the review file
-  crit unpublish                             Remove a shared review from crit-web
-  crit pull [--output <dir>] [pr-number]     Fetch GitHub PR comments into the review file
-  crit push [--dry-run] [--event <type>] [-m <msg>] [-o <dir>] [pr-number]  Post review comments to a GitHub PR
-  crit plan --name <slug> <file>             Review a plan file (manages versioned copies)
-  crit plan --name <slug>                    Read plan from stdin
-  crit auth login                            Log in to crit-web via browser
-  crit auth logout                           Log out and revoke token
-  crit auth whoami                           Show current user info
-  crit install <agent>                       Install integration files for an AI coding tool
-  crit status [--json]                        Print session info (review file, daemon, comments)
-  crit cleanup [--days N] [--force]           Delete stale review files (default: 7 days)
-  crit check                                 Check if installed integrations are up to date
-  crit config [--generate]                    Show resolved configuration
-  crit help                                  Show this help message
-
-  Agents:
-    claude-code, cursor, opencode, windsurf, github-copilot, cline, all
-
-Options:
-  -p, --port <port>           Port to listen on (default: random)
-  -o, --output <dir>          Output directory for review file
-      --no-open               Don't auto-open browser
-      --no-ignore             Disable all file ignore patterns
-  -q, --quiet                 Suppress status output
-      --share-url <url>       Share service URL (e.g. https://crit.md or self-hosted)
-      --base-branch <branch>  Base branch to diff against (overrides auto-detection)
-      --working-tree          Force working-tree mode (skip auto-detection of stacked PR / branch)
-      --scope <mode>          Diff scope when reviewing a PR: layer (default) or full-stack
-      --remote                Read PR file content via GitHub API instead of local git (requires gh)
-      --qr                    Print QR code of share URL (with crit share)
-  -v, --version               Print version
-
-Environment:
-  CRIT_SHARE_URL              Override the share service URL
-  CRIT_PORT                   Override the default port
-  CRIT_NO_UPDATE_CHECK        Disable update check on startup
-  CRIT_AUTH_TOKEN              Override the auth token (skip login)
-  CRIT_NO_INTEGRATION_CHECK   Disable integration staleness check
-
-Configuration:
-  Global config:   ~/.crit.config.json
-  Project config:  .crit.config.json (in repo root)
-  agent_cmd        Shell command to send comments to an AI agent (e.g. "claude -p")
-  Run 'crit config' to see resolved configuration.
-
-Learn more: https://crit.md
-`)
-}
-
-func printConfigHelp() {
-	fmt.Fprintf(os.Stderr, `crit config — show resolved configuration
-
-Prints the merged configuration from global and project config files as JSON.
-CLI flags and environment variables are not reflected in this output.
-
-Config files:
-  ~/.crit.config.json          Global config (applies to all projects)
-  .crit.config.json            Project config (in repo root)
-
-Precedence (highest to lowest):
-  1. CLI flags / env vars
-  2. Project config
-  3. Global config
-  4. Built-in defaults
-
-Available keys:
-  port              int       Port to listen on (default: random)
-  no_open           bool      Don't auto-open browser (default: false)
-  share_url         string    Share service URL
-  quiet             bool      Suppress status output (default: false)
-  output            string    Output directory for review file
-  author            string    Your name for comments (default: git config user.name)
-  base_branch       string    Base branch to diff against (overrides auto-detection)
-  ignore_patterns        []string  Gitignore-style patterns to exclude files from review
-  no_integration_check   bool      Skip integration staleness check (default: false)
-  agent_cmd              string    Shell command to send comments to an AI agent (e.g. "claude -p")
-  auth_token             string    Authentication token for crit-web share service
-
-Note: agent_cmd and auth_token are global-only (~/.crit.config.json).
-Project-level .crit.config.json cannot override them for security reasons.
-
-Ignore pattern syntax:
-  *.lock            Match files by extension (anywhere in tree)
-  vendor/           Match all files under a directory
-  package-lock.json Match exact filename (anywhere in tree)
-  generated/*.pb.go Match with path prefix (filepath.Match syntax)
-
-Example config:
-  {
-    "port": 3456,
-    "share_url": "https://crit.md",
-    "ignore_patterns": ["*.lock", "*.min.js", "vendor/", "generated/"]
-  }
-`)
-}
-
-func printVersion() {
-	line := "crit " + version
-	var details []string
-	if date != "unknown" {
-		details = append(details, date)
-	}
-	if commit != "unknown" {
-		short := commit
-		if len(short) > 7 {
-			short = short[:7]
-		}
-		details = append(details, short)
-	}
-	if len(details) > 0 {
-		line += " (" + strings.Join(details, ", ") + ")"
-	}
-	fmt.Println(line)
-	fmt.Println("Inline code review for AI agent workflows")
-}
-
-// globalDestKind selects how an integration's globalDest is interpreted.
-type globalDestKind int
-
-const (
-	// globalDestNone means the integration has no separate global path —
-	// use dest joined to home (the default for global installs).
-	globalDestNone globalDestKind = iota
-	// globalDestRelHome: globalDest is relative to $HOME.
-	globalDestRelHome
-	// globalDestDocuments: globalDest is relative to the platform Documents
-	// directory (used by Cline).
-	globalDestDocuments
-	// globalDestAbsolute: globalDest is an absolute path used verbatim.
-	globalDestAbsolute
-)
-
-type integration struct {
-	source string // path inside integrations/ embed
-	dest   string // destination relative to cwd
-	hint   string // usage hint printed after install
-	// globalDest, when set together with a non-zero globalDestKind, overrides
-	// dest in global mode (cwd == $HOME). The kind determines how it's
-	// resolved (see globalDestKind).
-	globalDest     string
-	globalDestKind globalDestKind
-}
-
-var integrationMap = map[string][]integration{
-	"claude-code": {
-		{source: "integrations/claude-code/skills/crit/SKILL.md", dest: ".claude/skills/crit/SKILL.md", hint: "Run /crit in Claude Code to start a review loop"},
-		{source: "integrations/claude-code/skills/crit-cli/SKILL.md", dest: ".claude/skills/crit-cli/SKILL.md", hint: "The crit-cli skill is available to Claude Code agents when needed"},
-	},
-	"cursor": {
-		{source: "integrations/cursor/skills/crit/SKILL.md", dest: ".cursor/skills/crit/SKILL.md", hint: "Run /crit in Cursor to start a review loop"},
-		{source: "integrations/cursor/skills/crit-cli/SKILL.md", dest: ".cursor/skills/crit-cli/SKILL.md", hint: "The crit-cli skill is available to Cursor agents when needed"},
-	},
-	"opencode": {
-		// command stays at ~/.opencode/commands/ globally (works there)
-		{source: "integrations/opencode/crit.md", dest: ".opencode/commands/crit.md", hint: "Run /crit in OpenCode to start a review loop"},
-		// opencode does NOT read ~/.opencode/skills/ globally — redirect to ~/.agents/skills/
-		{source: "integrations/opencode/SKILL.md", dest: ".opencode/skills/crit/SKILL.md", globalDest: ".agents/skills/crit/SKILL.md", globalDestKind: globalDestRelHome, hint: "The crit skill is available to OpenCode agents when needed"},
-	},
-	"windsurf": {
-		// windsurf has no per-tool global rules dir — global install rejected in installIntegration.
-		{source: "integrations/windsurf/crit.md", dest: ".windsurf/rules/crit.md", hint: "Windsurf will suggest Crit when writing plans"},
-	},
-	"github-copilot": {
-		// Copilot does NOT read ~/.github/skills/ globally — redirect to ~/.agents/skills/
-		{source: "integrations/github-copilot/skills/crit/SKILL.md", dest: ".github/skills/crit/SKILL.md", globalDest: ".agents/skills/crit/SKILL.md", globalDestKind: globalDestRelHome, hint: "Run /crit in GitHub Copilot to start a review loop"},
-		{source: "integrations/github-copilot/skills/crit-cli/SKILL.md", dest: ".github/skills/crit-cli/SKILL.md", globalDest: ".agents/skills/crit-cli/SKILL.md", globalDestKind: globalDestRelHome, hint: "The crit-cli skill is available to GitHub Copilot agents when needed"},
-	},
-	"cline": {
-		// Cline does NOT read ~/.clinerules/ globally — redirect to platform Documents dir.
-		{source: "integrations/cline/crit.md", dest: ".clinerules/crit.md", globalDest: "Cline/Rules/crit.md", globalDestKind: globalDestDocuments, hint: "Cline will suggest Crit when writing plans"},
-	},
-	"codex": {
-		{source: "integrations/codex/skills/crit/SKILL.md", dest: ".agents/skills/crit/SKILL.md", hint: "Use $crit in Codex to start a review loop"},
-		{source: "integrations/codex/skills/crit-cli/SKILL.md", dest: ".agents/skills/crit-cli/SKILL.md", hint: "The crit-cli skill is available to Codex agents when needed"},
-	},
-}
-
-// availableIntegrations returns the sorted list of integration names that
-// `crit install <name>` accepts. Derived from integrationMap keys plus the
-// special-cased "aider" entry (which does not live in the map because its
-// install flow is bespoke — see installAider).
-func availableIntegrations() []string {
-	names := make([]string, 0, len(integrationMap)+1)
-	for name := range integrationMap {
-		names = append(names, name)
-	}
-	names = append(names, "aider")
-	sort.Strings(names)
-	return names
-}
-
-// isGlobalInstall reports whether the install should be treated as global
-// (user-wide) rather than project-scoped. True when cwd == $HOME.
-func isGlobalInstall(cwd, home string) bool {
-	if cwd == "" || home == "" {
-		return false
-	}
-	a, errA := filepath.Abs(cwd)
-	b, errB := filepath.Abs(home)
-	if errA != nil || errB != nil {
-		return cwd == home
-	}
-	return a == b
-}
-
-// resolveGlobalDest expands an integration's globalDest into an absolute
-// path according to its globalDestKind.
-func resolveGlobalDest(kind globalDestKind, globalDest, home string) (string, error) {
-	switch kind {
-	case globalDestAbsolute:
-		return globalDest, nil
-	case globalDestDocuments:
-		return filepath.Join(documentsDir(home), globalDest), nil
-	case globalDestRelHome, globalDestNone:
-		if filepath.IsAbs(globalDest) {
-			return globalDest, nil
-		}
-		return filepath.Join(home, globalDest), nil
-	default:
-		return "", fmt.Errorf("unknown globalDestKind %d", kind)
-	}
-}
-
-// xdgUserDirFn is the seam used by documentsDir to query xdg-user-dir.
-// Tests override this; production code uses the default that shells out.
-var xdgUserDirFn = xdgUserDir
-
-// documentsDir returns the platform Documents directory for the current user.
-//
-//	macOS:   $HOME/Documents
-//	Linux:   $(xdg-user-dir DOCUMENTS), falling back to $HOME/Documents.
-//	         If xdg-user-dir returns $HOME (its documented behavior when
-//	         user-dirs.dirs is missing), we treat that as "no answer" and
-//	         fall back to $HOME/Documents to avoid polluting the home dir.
-//	Windows: filepath.Join(home, "Documents") — the real MyDocuments folder
-//	         can differ; querying FOLDERID_Documents needs x/sys/windows.
-//	         The $USERPROFILE\Documents convention is a pragmatic default.
-func documentsDir(home string) string {
-	if runtime.GOOS == "linux" {
-		path, err := xdgUserDirFn("DOCUMENTS")
-		if err == nil && path != "" && path != home {
-			return path
-		}
-	}
-	return filepath.Join(home, "Documents")
-}
-
-// xdgUserDir shells out to the xdg-user-dir binary to query a user dir.
-// Returns ("", error) if the binary is missing or returns non-zero. The
-// returned path is whitespace-trimmed; it may equal $HOME when the spec
-// says no user-dirs.dirs entry exists — callers must handle that case.
-func xdgUserDir(name string) (string, error) {
-	out, err := exec.Command("xdg-user-dir", name).Output()
-	if err != nil {
-		return "", err
-	}
-	return strings.TrimSpace(string(out)), nil
-}
-
-// installIntegration installs the named agent integration. It returns an
-// error suitable for printing to stderr; callers decide whether to exit or
-// continue (the `install all` loop continues past per-agent failures).
-func installIntegration(name string, force bool) error {
-	if name == "aider" {
-		return installAider(force)
-	}
-
-	files, ok := integrationMap[name]
-	if !ok {
-		var b strings.Builder
-		fmt.Fprintf(&b, "unknown agent: %s\n\nAvailable agents:\n", name)
-		for _, a := range availableIntegrations() {
-			fmt.Fprintf(&b, "  %s\n", a)
-		}
-		return errors.New(strings.TrimRight(b.String(), "\n"))
-	}
-
-	cwd := mustGetwd()
-	home, _ := os.UserHomeDir()
-	global := isGlobalInstall(cwd, home)
-
-	if name == "windsurf" && global {
-		return errors.New("windsurf does not support a global per-tool install. " +
-			"Windsurf only loads a single ~/.codeium/windsurf/memories/global_rules.md (6k char cap), " +
-			"not a per-tool rules directory. Run `crit install windsurf` from a project directory " +
-			"instead, which writes .windsurf/rules/crit.md (workspace-scoped)")
-	}
-
-	var hints []string
-	for _, f := range files {
-		dest := destFor(f, global, home, name)
-		installOneFile(f, dest, force)
-		if f.hint != "" {
-			hints = append(hints, f.hint)
-		}
-	}
-	printUniqueHints(hints)
-	fmt.Println()
-	return nil
-}
-
-// destFor returns the destination path for an integration file, accounting
-// for global vs project install mode.
-func destFor(f integration, global bool, home, name string) string {
-	if global && f.globalDest != "" {
-		resolved, err := resolveGlobalDest(f.globalDestKind, f.globalDest, home)
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "Error resolving global destination for %s: %v\n", name, err)
-			os.Exit(1)
-		}
-		return resolved
-	}
-	return f.dest
-}
-
-// installOneFile copies a single embedded integration file to dest, skipping
-// if it already exists (unless force is set). Exits on I/O errors.
-func installOneFile(f integration, dest string, force bool) {
-	if !force {
-		if _, err := os.Stat(dest); err == nil {
-			fmt.Printf("  Skipped:   %s (already exists, use --force to overwrite)\n", dest)
-			return
-		}
-	}
-	data, err := integrationsFS.ReadFile(f.source)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error reading embedded file %s: %v\n", f.source, err)
-		os.Exit(1)
-	}
-	if err := os.MkdirAll(filepath.Dir(dest), 0o755); err != nil {
-		fmt.Fprintf(os.Stderr, "Error creating directory %s: %v\n", filepath.Dir(dest), err)
-		os.Exit(1)
-	}
-	if err := os.WriteFile(dest, data, 0o644); err != nil {
-		fmt.Fprintf(os.Stderr, "Error writing %s: %v\n", dest, err)
-		os.Exit(1)
-	}
-	fmt.Printf("  Installed: %s\n", dest)
-}
-
-// printUniqueHints prints each hint once, in the order it first appeared.
-func printUniqueHints(hints []string) {
-	seen := make(map[string]bool)
-	for _, hint := range hints {
-		if seen[hint] {
-			continue
-		}
-		seen[hint] = true
-		fmt.Printf("  %s\n", hint)
-	}
 }
